Splitting user class
