+++ backend/db/src/main/java/com/bls/patronage/db/model/User.java
@@ -4,31 +4,17 @@ import com.google.common.base.Objects;
 
 import java.util.UUID;
 
-public class User extends IdentifiableEntity {
-    private String email;
-    private String name;
+public class User extends UserWithoutId {
     private String password;
 
     public User(String id, String email, String name, String password) {
-        super(id);
-        this.email = email;
-        this.name = name;
+        super(id, email, name);
         this.password = password;
     }
 
     public User(UUID id, String email, String name, String password) {
-        super(id);
-        this.email = email;
-        this.name = name;
+        super(id, email, name);
         this.password = password;
-    }
-
-    public String getEmail() {
-        return email;
-    }
-
-    public String getName() {
-        return name;
     }
 
     public String getPassword() {
@@ -37,13 +23,11 @@ public class User extends IdentifiableEntity {
 
     @Override
     public boolean equals(Object o) {
-        if (this == o) return true;
-        if (o == null || getClass() != o.getClass()) return false;
-        User user = (User) o;
-        return Objects.equal(getId(), user.getId()) &&
-                Objects.equal(getEmail(), user.getEmail()) &&
-                Objects.equal(getName(), user.getName()) &&
-                Objects.equal(getPassword(), user.getPassword());
+        if (super.equals(o)) {
+            return Objects.equal(getPassword(), ((User) o).getPassword());
+        } else {
+            return false;
+        }
     }
 
     @Override
